remove float timestamp

+++ socket_server/server.py
@@ -68,7 +68,7 @@ class ChatCenter(object):
 
     @staticmethod
     def get_now_end_time(duration):
-        now = float(time.time())
+        now = int(time.time())
         return now + duration
 
     def response_wrapper(self, message, status=STATUS_SUCCESS, msg='success', msg_type=1, raw_message=None):
@@ -282,7 +282,7 @@ class ChatCenter(object):
             yield sender.write_message(self.response_wrapper({}, STATUS_ERROR, '不能演唱不是自己点的歌~', raw_message=message))
             return
         song = self.songs.pop(message.room)
-        song['duration'] = float(song.get('duration'), 0)
+        song['duration'] = int(song.get('duration'), 0)
         self.user_song.remove_member_from_set(message.room, message.fullname)
         if ack:
             # celery task id
@@ -293,7 +293,7 @@ class ChatCenter(object):
             yield self.boardcast_in_room(sender, res)
             # 歌曲完成回调
             singing_callback.apply_async((message.room, res.get('end_time'), task, res.get('duration')),
-                                         countdown=float(res.get('duration')))
+                                         countdown=int(res.get('duration')))
         else:
             song = self.songs.get(message.room)
             if not song:
@@ -335,7 +335,7 @@ class ChatCenter(object):
         music = self.music.pop(message.room)
         if music:
             self.user_music.remove_member_from_set(message.room, message.fullname)
-            duration = float(music.get('duration'))
+            duration = int(music.get('duration'))
             res = self.room.set_music(message.room, music, task)
             music_callback.apply_async((message.room, self.get_now_end_time(duration), task, duration),
                                        countdown=duration)
@@ -432,7 +432,7 @@ class ChatCenter(object):
             song = self.music.pop(message.room)
             self.user_music.remove_member_from_set(message.room, message.fullname)
             task = generate_task_id()
-            duration = float(song.get('duration'))
+            duration = int(song.get('duration'))
             res = self.room.set_music(message.room, song, task)
             room_status = self.get_room_info(message.room)
             music_callback.apply_async((message.room, self.get_now_end_time(duration), task, duration),
